Flow Sensor for Drainage System to detect blockage
Drainage is defined as the infrastructure for drying the land from the excess and unutilized water; rainwater and waste water. Drainage conditions should be monitored in order to maintain its proper function. In fact, not all areas have drainage monitoring team. It leads to irregular monitoring of the drainage condition. The irregular monitoring has contribution on the clogging of the drainage that imply to the siltation which trigger flooding in the neighbourhood. Manual monitoring is also inefficient. It needs a lot of dedicated persons who are only able to record limited report with low accuracy. These weaknesses lead to the slow handling for problems in drainage. This work discuss about the design of drainage systems intended to monitor conditions at several points in drainage system using wireless sensor networks. Some sensor nodes are deployed at several determined points to be connected each. The monitoring parameters are nothing but the flow of water through the drainage pipe lines. This system is expected to monitor the drainage conditions real time continuously. If the flow is decreases it is alert that the flow is decreases ,then we can come to know that there is a blockage..